[image] Rename `convertion*` files into `conversion*`
