+++ src/aliceVision/image/all.hpp
@@ -14,7 +14,7 @@
 
 #include "aliceVision/image/Image.hpp"
 #include "aliceVision/image/pixelTypes.hpp"
-#include "aliceVision/image/convertion.hpp"
+#include "aliceVision/image/conversion.hpp"
 #include "aliceVision/image/drawing.hpp"
 #include "aliceVision/image/filtering.hpp"
 #include "aliceVision/image/resampling.hpp"
@@ -26,6 +26,6 @@
 #include "aliceVision/image/convolution.hpp"
 #include "aliceVision/image/Rgb.hpp"
 #include "aliceVision/image/Sampler.hpp"
-#include "aliceVision/image/convertionOpenCV.hpp"
+#include "aliceVision/image/conversionOpenCV.hpp"
 #include "aliceVision/image/dcp.hpp"
 #include "aliceVision/image/ImageCache.hpp"
