+++ src/aliceVision/image/conversion.hpp
