+++ src/software/export/main_exportMatlab.cpp
@@ -7,7 +7,7 @@
 #include <aliceVision/sfmData/SfMData.hpp>
 #include <aliceVision/sfmDataIO/sfmDataIO.hpp>
 #include <aliceVision/image/all.hpp>
-#include <aliceVision/image/convertion.hpp>
+#include <aliceVision/image/conversion.hpp>
 #include <aliceVision/system/main.hpp>
 #include <aliceVision/cmdline/cmdline.hpp>
 #include <boost/program_options.hpp>
